+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * fixed selecting largest icon
 * fixed fallback to /favicon.ico if index cannot be loaded
+* fixed grouping sizes
 
 ## 1.0.0